+++ libImageIO/src/imageio/FileImageSource.cpp
@@ -57,19 +57,19 @@ void FileImageSource::reset()
 bool FileImageSource::next()
 {
 	index++;
-	return index < files.size();
+	return index < static_cast<int>(files.size());
 }
 
 const Mat FileImageSource::getImage() const
 {
-	if (index < 0 || index >= files.size())
+	if (index < 0 || index >= static_cast<int>(files.size()))
 		return Mat();
 	return imread(files[index].string(), 1);
 }
 
 path FileImageSource::getName() const
 {
-	if (index < 0 || index >= files.size())
+	if (index < 0 || index >= static_cast<int>(files.size()))
 		return path();
 	return files[index];
 }
