+++ libImageIO/src/imageio/FileListImageSource.cpp
@@ -71,19 +71,19 @@ void FileListImageSource::reset()
 bool FileListImageSource::next()
 {
 	index++;
-	return index < files.size();
+	return index < static_cast<int>(files.size());
 }
 
 const Mat FileListImageSource::getImage() const
 {
-	if (index < 0 || index >= files.size())
+	if (index < 0 || index >= static_cast<int>(files.size()))
 		return Mat();
 	return imread(files[index].string(), 1);
 }
 
 path FileListImageSource::getName() const
 {
-	if (index < 0 || index >= files.size())
+	if (index < 0 || index >= static_cast<int>(files.size()))
 		return path();
 	return files[index];
 }
